pyproject.toml
[project]
name = "event-driven-app-using-nats"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "nats-py>=2.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
